requirements.txt
flask
fiona
geopy
utm
pyhton-telegram-bot
requests
staticmap
